Convert to TypeScript C-F
